refactoring and fix

+++ Warehouse/Models/Dtos/CreateCoilDto.cs
@@ -4,9 +4,9 @@ namespace Warehouse.Models.Dtos;
 
 public class CreateCoilDto
 {
-    [Range(0, int.MaxValue, ErrorMessage = "Lenght must be non-negative")]
-    public required int Lenght { get; set; } 
+    [Range(0, long.MaxValue, ErrorMessage = "Lenght must be non-negative")]
+    public required long Lenght { get; set; } 
     
-    [Range(0, int.MaxValue, ErrorMessage = "Weight must be non-negative")]
-    public required int Weight { get; set; }
+    [Range(0, long.MaxValue, ErrorMessage = "Weight must be non-negative")]
+    public required long Weight { get; set; }
 }
